requirements.txt
# apscheduler==3.2.0

# Waiting on https://github.com/agronholm/apscheduler/pull/149
apscheduler-ng